% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund_to_recruit.R
\name{abund2recruit}
\alias{abund2recruit}
\title{Convert abundance estimates to a daily recruitment profile
for one region/year}
\usage{
abund2recruit(abund, phi = 0.6)
}
\arguments{
\item{abund}{An abundance record as returned from clean_data}

\item{phi=0.6}{}
}
\value{
a dataframe with estimated daily abundance profile.  Note that these
are not trimmed to nonnegative values.
}
\description{
Convert abundance estimates to a daily recruitment profile
for one region/year
}
\examples{
library(readxl)
abund <- clean_data(read_excel("annualabundances/2015.xls"))
upu <- filter(abund,region=="Uncompahgre Upper")
abund2recruit(upu)
qplot(date,recruit,data=abund2recruit(upu))
all_recruit_profile <- abund \%>\% group_by(region) \%>\% do(abund2recruit(.))
qplot(date,recruit,data=all_recruit_profile,color=region,geom="line")
}

