% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund_to_recruit.R
\name{boot_ci}
\alias{boot_ci}
\title{Calculate a 95% confidence interval for a single region/year using
bootstrapping.}
\usage{
boot_ci(nboot, abund, phi = 0.6)
}
\arguments{
\item{nboot}{number of replications (typically 1000)}

\item{abund}{an abundance record for one region/year}

\item{phi=0.6}{}
}
\value{
a dataframe with the estimate and upper and lower confidence
limits
}
\description{
Calculate a 95% confidence interval for a single region/year using
bootstrapping.
}
\examples{
abund <- clean_data(read_excel("annualabundances/2015.xls"))
upu <- filter(abund,region=="Uncompahgre Upper")
boot_ci(1000,upu)
abund \%>\% group_by(region) \%>\% do(boot_ci(1000,.))
}

