% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_clean.R
\name{clean_data}
\alias{clean_data}
\title{clean_data prepares data from Distance software for analysis}
\usage{
clean_data(df)
}
\arguments{
\item{df}{A dataframe obtained from read_excel on an annual abundance
file.  The format for this file is: date region sub_region est lower upper
where date is in ISO: YYYY-MM-DD, region is main region (title case) and
sub_region is title case.}
}
\value{
A cleaned version with the dates converted from integers
and a single region column.  Note that region and sub_region should
be in title case.
}
\description{
clean_data prepares data from Distance software for analysis
}
\examples{
library(readxl)
clean_data(read_excel("filename"))
}

