% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_clean.R
\name{format_for_distance}
\alias{format_for_distance}
\title{Format a data frame produced by read_ufb_datasheet for importing
into distance software}
\usage{
format_for_distance(dat)
}
\arguments{
\item{dat}{}
}
\value{
a data frame with appropriate naming and date conventions for
distance software
}
\description{
Format a data frame produced by read_ufb_datasheet for importing
into distance software
}
\examples{
library(readxl)
ds <- read_ufb_datasheet("Distance Input 2015.xlsx")
dsf <- format_for_distance(ds)
write.csv(dsf,row.names=FALSE,na="",file="distanceinput2015.csv", eol = "\\r\\n")
}

