% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund_to_recruit.R
\name{makebootstrap}
\alias{makebootstrap}
\title{Make a bootstrap sample from an abundance record for one region/year}
\usage{
makebootstrap(abund)
}
\arguments{
\item{abund}{A dataframe with columns est, lower, upper}
}
\value{
a similarly formatted dataframe with estimates varied by
log normal distributions based on the original estimate and upper
confidence level
}
\description{
Make a bootstrap sample from an abundance record for one region/year
}
\examples{
abund <- clean_data(read_excel("annualabundances/2015.xls"))
upu <- filter(abund,region=="Uncompahgre Upper")
makebootstrap(upu)
apply(replicate(1000, makebootstrap(upu)$est),1,
 function(dat)quantile(dat,probs=c(0.025,0.975)))
}

