% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_clean.R
\name{read_ufb_datasheet}
\alias{read_ufb_datasheet}
\title{Take an excel file recording annual observations and build a data frame
appropriate for converting to distance software}
\usage{
read_ufb_datasheet(filename)
}
\arguments{
\item{filename}{}
}
\value{
A data frame with appropriate columns
}
\description{
Take an excel file recording annual observations and build a data frame
appropriate for converting to distance software
}
\examples{

library(readxl)
read_ufb_datasheet("Distance Input 2015.xlsx")
write.csv(di,row.names=FALSE,na="",file="distanceinput2015.csv", eol = "\\r\\n")
}

