% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund_to_recruit.R
\name{total_recruit}
\alias{total_recruit}
\title{Calculate total recruitment for one region/year}
\usage{
total_recruit(abund, phi = 0.6)
}
\arguments{
\item{abund}{An abundance record as returned from clean_data}

\item{phi=0.6}{}
}
\value{
The sum of the nonnegative recruitments
}
\description{
Calculate total recruitment for one region/year
}
\examples{
abund <- clean_data(read_excel("annualabundances/2015.xls"))
upu <- filter(abund,region=="Uncompahgre Upper")
total_recruit(upu)
all_recruit <- abund \%>\% group_by(region) \%>\% do(data_frame(total=total_recruit(.)))
}

